python_basic_diploma-master/python_basic_diploma-master/requirements.txt
python-dotenv==1.0.1
PyTelegramBotAPI==4.16.1
requests==2.31.0